refactor: UserRepository 이동 #50
- auth module에 있는 UserRepository를 users module로 이동
- users module에서 UserRepository export하고 auth module에서 import
