+++ backend/src/auth/auth.module.ts
@@ -2,14 +2,15 @@ import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { NaverStrategy } from './strategy/naverAuth.strategy';
-import { UserRepository } from './auth.repository';
 import { JwtModule } from '@nestjs/jwt';
 import 'dotenv/config';
 import { JwtAuthStrategy } from './strategy/jwtAuth.strategy';
 import { JWT_EXPIRE_DATE } from './utils/auth.constant';
+import { UsersModule } from 'src/users/users.module';
 
 @Module({
   imports: [
+    UsersModule,
     JwtModule.register({
       secret: process.env.JWT_SECRET,
       signOptions: {
@@ -18,6 +19,6 @@ import { JWT_EXPIRE_DATE } from './utils/auth.constant';
     }),
   ],
   controllers: [AuthController],
-  providers: [AuthService, UserRepository, NaverStrategy, JwtAuthStrategy],
+  providers: [AuthService, NaverStrategy, JwtAuthStrategy],
 })
 export class AuthModule {}
