+++ backend/src/auth/guards/naverAuth.guard.ts
@@ -1,6 +1,6 @@
 import { AuthGuard } from '@nestjs/passport';
 import { Injectable } from '@nestjs/common';
-import { SocialType } from '../entity/socialType';
+import { SocialType } from '../../users/entity/socialType';
 
 @Injectable()
 export class NaverAuthGuard extends AuthGuard(SocialType.NAVER) {}
