+++ backend/src/auth/strategy/jwtAuth.strategy.ts
@@ -3,7 +3,7 @@ import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import 'dotenv/config';
 import { VerifyCallback } from '../utils/verifyCallback';
-import { UserRepository } from '../auth.repository';
+import { UserRepository } from '../../users/auth.repository';
 import { Request } from 'express';
 import { JWT, Payload } from '../utils/jwt.type';
