+++ backend/src/users/users.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
+import { UserRepository } from './auth.repository';
 
 @Module({
   controllers: [UsersController],
-  providers: [UsersService],
+  providers: [UsersService, UserRepository],
+  exports: [UserRepository],
 })
 export class UsersModule {}
